sc-meta tx - file rename
