+++ framework/meta/src/cmd/tx/tx_cli_sign.rs
@@ -4,7 +4,7 @@ use multiversx_sc_snippets::{hex, sdk::utils::base64_decode};
 use super::{
     output::TxOutputFile,
     tx_cli_args::SignArgs,
-    tx_common::{broadcast_and_save, load_transaction_from_file, load_wallet, save_output},
+    tx_cli_common::{broadcast_and_save, load_transaction_from_file, load_wallet, save_output},
 };
 
 pub async fn tx_sign(args: &SignArgs) {
